+++ src/middlewares/body-parser.ts
@@ -8,7 +8,7 @@ import Formidable from "formidable";
 
 import type { FritterContext } from "../classes/FritterContext.js";
 
-import type { FritterMiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
+import type { MiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
 import type { PossibleJsonData } from "../types/PossibleJsonData.js";
 
 //
@@ -52,7 +52,7 @@ export interface CreateOptions
 
 export interface CreateResult
 {
-	execute : FritterMiddlewareFunction<MiddlewareFritterContext>;
+	execute : MiddlewareFunction<MiddlewareFritterContext>;
 }
 
 export function create(options? : CreateOptions) : CreateResult