+++ src/middlewares/force-ssl.ts
@@ -6,7 +6,7 @@ import { isLocalIpAddress } from "@donutteam/utilities";
 
 import { FritterContext } from "../classes/FritterContext.js";
 
-import type { FritterMiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
+import type { MiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
 
 //
 // Interfaces
@@ -28,7 +28,7 @@ export interface CreateOptions
 
 export interface CreateResult
 {
-	execute : FritterMiddlewareFunction<MiddlewareFritterContext>;
+	execute : MiddlewareFunction<MiddlewareFritterContext>;
 }
 
 export function create(options? : CreateOptions) : CreateResult