+++ src/middlewares/router.ts
@@ -10,7 +10,7 @@ import { pathToRegexp, type Key, type ParseOptions, type TokensToRegexpOptions }
 
 import type { FritterContext } from "../classes/FritterContext.js";
 
-import type { FritterMiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
+import type { MiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
 import type { HTTPMethod } from "../types/HTTPMethod.js";
 
 //
@@ -46,7 +46,7 @@ export interface CreateOptions
 
 export interface CreateResult
 {
-	execute : FritterMiddlewareFunction<MiddlewareFritterContext>;
+	execute : MiddlewareFunction<MiddlewareFritterContext>;
 
 	getRoutes : () => Route[];