+++ src/types/FritterMiddlewareFunction.ts
@@ -8,4 +8,7 @@ import { FritterContext } from "../classes/FritterContext.js";
 // Types
 //
 
-export type FritterMiddlewareFunction<Type extends FritterContext = FritterContext> = (context : Type, next : () => Promise<void>) => Promise<void>;
+export type MiddlewareFunction<Type extends FritterContext = FritterContext> = (context : Type, next : () => Promise<void>) => Promise<void>;
+
+/** @deprecated */
+export type FritterMiddlewareFunction<Type extends FritterContext = FritterContext> = MiddlewareFunction<Type>;